unisim/kernel/service: 	+ just rewriten VariableBase () operator for knowing implicit the cast it is doing.
unisim/kernel/inline_debugger:
	+ added support for monitoring services variables.
	- TODO: add help for the three parameters version of the monitor command (monitor with explicit casting)
	- TODO: replace explicit casting of three parameters monitor by something more gdb like

unisim/component/cxx/processor/tms320:
	+ fixed update of AR register on instructions of type indir_reg and reg_indir
	


git-svn-id: https://unisim.org/svn/devel@11237 ea533459-3c0a-0410-a024-a90cf4b01ff5

+++ unisim_lib/unisim/kernel/service/service.cc
@@ -195,7 +195,7 @@ template <class TYPE> Variable<TYPE>::operator bool () const { return (*storage)
 template <class TYPE> Variable<TYPE>::operator long long () const { return (long long) *storage; }
 template <class TYPE> Variable<TYPE>::operator unsigned long long () const { return (unsigned long long) *storage; }
 template <class TYPE> Variable<TYPE>::operator double () const { return (double) *storage; }
-template <class TYPE> Variable<TYPE>::operator string () const { stringstream sstr; sstr << "0x" << hex << *storage; return sstr.str(); }
+template <class TYPE> Variable<TYPE>::operator string () const { stringstream sstr; sstr << "0x" << hex << (unsigned long long) *storage; return sstr.str(); }
 
 template <class TYPE> VariableBase& Variable<TYPE>::operator = (bool value) { *storage = value ? 1 : 0; return *this; }
 template <class TYPE> VariableBase& Variable<TYPE>::operator = (long long value) { *storage = value;	return *this; }
